requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
streamlit
Pillow